Reorganize auth plugins
